Separate network name and url
This makes the name of a network and the url used separate. Ex. You can
have a network named 'fn' which connects to the url irc.freenode.net.
This will be more useful when configuration file support is added. The
name, (Ex. 'fn') is what is used when creating the directory.