Add configuration file and configurartion support
This patch modifies fircd to use libconfuse for config file support. The
default configuration file is ~/.fircdrc. This also removes a lot of the
old command-line arguments, as the configuration file replaces their
use. The configuration file syntax is documented in ./doc/fircd.1 and an
example config file is included as ./doc/fircdrc.example.