Changes to channel file-system.
This patch adds two new files to channel, 'msgs' and 'raw'. 'msgs' is
the combo of all the messages sent to the channel (Lines sent to 'out')
without any extra information, Ex. topic changes, nick changes, etc.
'raw' is intended for client use, as a simple to parse file-format. With
all information on what's happening to the channel.