Refactor channel organization
This change removes the linked-list node from the channel struct, and
instead has the channel struct get wrapped inside of another struct
containing a node at the network level. The advantage here is that the
organization of the channel's is now at the network level and it results
in cleaner code.