Refactoring config setup
This patch refactors the config setup in ./include/config.h and
./src/config.c. Specefically it moves everything config related out of
struct network_cons and into struct config, and cleans up parts of the
code. It also makes the config global over the program, which simplifies
a ton of the code (Before they had to et sent network_cons just to use
config options in it).