Refactor libconfuse and add was_set option
This patch refactors libconfuse to remove some unneeded options
(DLLEXPORT, __export, etc...), as well as adds a 'was_set' option, which
flags whether or not an option was actually defined and set in the
config file, or if the default was used. Previous this functionality was
acheved via hacking around how libconfuse worked, but this is a
nessisary change to allow for having global defaults that are
overridable on a per-network basis.