Fixing confuse.c memory leak and access issues
The first leak is caused via not freeing the opttitle variable on a
return (Fixed via removing all returns and turning them into jumps to a
clean-up at the end which free's the variable if nessisary).

The second access issue has to do with assigning a linked-list directly
to internal sections, and then freeing it for each section (Resulting in
a double free on the list). The easy solution is to simply make a
complete copy of the list whenever a section is loaded, so no double
free happens.