Include macros for dprintf as well as vararg alloc_sprintf
This is two small changes. The first is the inclusion of two 'fdprintf'
macros, which do a printf to a file descriptor. At the current moment,
that works via mapping it directly to dprintf, however that could change
for compatibility reasons later-on. The second change is adding the
'alloc_sprintfv' function.