Refactor channel.c to use dfprintf instead of alloc_sprintf
This refactor changes out all the allocations used for alloc_sprintf
plus write(), and instead uses the 'dfprintf' macro to handle it (Which
maps to dprintf).

Eventually this will be replaced too, with a more customizable safer
implementation of a printf type syntax, which could then be expressed in
the configuration files to allow these messages to be configured.